scroll_text: render frames in advance
Previously, the script just waited for 1/FPS s before rendering the next
frame and immediately displaying it. So the assumption is that rendering
the next frame takes zero seconds. This however is not even close enough,
since the script is deployed on a Raspberry Pi for flipdot-gschichtler,
where rendering a frame sometimes takes a bit longer. This leads to very
unpleasant frame rate incosistencies.

This is fixed by having one thread fill the frame buffer and another one
sending the next frame every 1/FPS seconds.